# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  service.BindDeviceRequest:
    properties:
      device_id:
        type: string
      ip_address:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
    required:
    - device_id
    - user_id
    type: object
...
  service.ValidateDeviceRequest:
    properties:
      bind_token:
        type: string
      device_id:
        type: string
      ip_address:
        type: string
      user_id:
        type: string
    required:
    - bind_token
    - device_id
    - user_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: Auth Service for B2B Suite – admin, employee, attendance, etc.
  title: B2B Suite API
  version: "1.0"
paths:
...
  /api/v1/devices/{userID}:
    delete:
      parameters:
      - description: User UUID
        in: path
        name: userID
        required: true
        type: string
      responses:
        "200":
          description: Device unbound
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Unbind failed
          schema:
            additionalProperties: true
            type: object
      summary: Unbind device
      tags:
      - devices
    get:
      parameters:
      - description: User UUID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Active device
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No active device
          schema:
            additionalProperties: true
            type: object
      summary: Get active device
      tags:
      - devices
  /api/v1/devices/{userID}/history:
    get:
      parameters:
      - description: User UUID
        in: path
        name: userID
        required: true
        type: string
      - default: 100
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: History list
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
      summary: Get device history
      tags:
      - devices
  /api/v1/devices/bind:
    post:
      consumes:
      - application/json
      description: Associates a device ID with a user.
      parameters:
      - description: Bind request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.BindDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Device bound
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Binding failed
          schema:
            additionalProperties: true
            type: object
      summary: Bind device
      tags:
      - devices
  /api/v1/devices/cleanup:
    post:
      parameters:
      - default: 90
        description: Age in days
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cleanup result
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Cleanup failed
          schema:
            additionalProperties: true
            type: object
      summary: Cleanup orphaned devices
      tags:
      - devices
...
  /api/v1/devices/search:
    get:
      parameters:
      - description: Device ID
        in: query
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing device_id
          schema:
            additionalProperties: true
            type: object
      summary: Get users by device
      tags:
      - devices
  /api/v1/devices/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Statistics
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get stats
          schema:
            additionalProperties: true
            type: object
      summary: Get device service stats
      tags:
      - devices
  /api/v1/devices/validate:
    post:
      consumes:
      - application/json
      description: Checks if a device is trusted and valid.
      parameters:
      - description: Validation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ValidateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Validation result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Validation failed
          schema:
            additionalProperties: true
            type: object
      summary: Validate device
      tags:
      - devices
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"